# architecture/openapi.yaml
openapi: 3.0.3
info:
  title: Klawa API
  description: API для личного ИИ-помощника Klawa по управлению задачами.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Основной сервер
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TaskStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, COMPLETED, ARCHIVED]
      description: Статус задачи
    TaskPriority:
      type: string
      enum: [LOW, MEDIUM, HIGH, URGENT]
      description: Приоритет задачи
...
    UserRole:
      type: string
      enum: [ROLE_USER, ROLE_ADMIN]
      description: Роль пользователя
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RegisterRequest:
      type: object
      required: [email, password, firstName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        roles:
          $ref: '#/components/schemas/UserRole'
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    ProjectResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        userId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectPagedResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer

    # -----------------------------
    # DTOs: Tasks
    # -----------------------------
...
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        projectId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        dueDate:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskPagedResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          
    # -----------------------------
    # DTOs: AI / Conversations
    # -----------------------------
...
paths:
...
  /auth/authenticate:
    post:
      tags: [Auth]
      summary: Аутентификация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Неверные данные
        '401':
          description: Неверный логин или пароль
  /auth/register:
    post:
      tags: [Auth]
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Ошибка валидации
        '409':
          description: Email уже занят
  /auth/me:
    get:
      tags: [Auth]
      summary: Получение профиля текущего пользователя
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Не авторизован
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Обновление токена доступа
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Токен успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Токен недействителен
  /auth/logout:
    post:
      tags: [Auth]
      summary: Выход пользователя (отзыв токена)
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Успешный выход
        '401':
          description: Не авторизован

  # ==========================================
  # USERS
  # ==========================================
...
  /projects:
    get:
      tags: [Projects]
      summary: Получение списка проектов
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список проектов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPagedResponse'
        '401':
          description: Не авторизован

    post:
      tags: [Projects]
      summary: Создание нового проекта
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
      responses:
        '201':
          description: Проект создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Ошибка валидации
        '401':
          description: Не авторизован
  /projects/{id}:
    get:
      tags: [Projects]
      summary: Получение проекта по ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Детали проекта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '401':
          description: Не авторизован
        '404':
          description: Проект не найден

    put:
      tags: [Projects]
      summary: Обновление проекта
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
      responses:
        '200':
          description: Проект обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '401':
          description: Не авторизован
        '404':
          description: Проект не найден

    delete:
      tags: [Projects]
      summary: Удаление проекта
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Проект удален
        '401':
          description: Не авторизован
        '404':
          description: Проект не найден
  /projects/{id}/tasks:
    get:
      tags: [Projects]
      summary: Получение списка задач для проекта
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список задач
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskPagedResponse'
        '401':
          description: Не авторизован
        '404':
          description: Проект не найден

  # ==========================================
  # AGENT / AI
  # ==========================================
...